requirements.txt
# Core LangChain dependencies
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0

# Environment and configuration
python-dotenv>=1.0.0

# Web framework dependencies
streamlit>=1.28.0
flask>=2.3.0
flask-cors>=4.0.0

# HTTP and API clients
requests>=2.31.0
httpx>=0.24.0

# Data handling
pydantic>=2.0.0

# Optional: For enhanced functionality
gradio>=3.50.0

# Development and testing (optional)
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.25.0

# Security and utilities
cryptography>=41.0.0